requirements.txt
google-adk[database]==0.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
firebase-admin==6.5.0
fastapi
opencv-python==4.10.0.84
deprecated